requirements.txt
Django==5.0.2
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1 